+++ app/src/main/java/com/techcourse/UncheckedServletException.java
@@ -1,0 +1,8 @@
+package com.techcourse;
+
+public class UncheckedServletException extends RuntimeException {
+
+    public UncheckedServletException(Exception e) {
+        super(e);
+    }
+}
