+++ app/src/main/java/com/techcourse/controller/RegisterController.java
@@ -4,7 +4,7 @@ import com.techcourse.domain.User;
 import com.techcourse.repository.InMemoryUserRepository;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.controller.asis.Controller;
+import webmvc.org.springframework.web.servlet.mvc.asis.Controller;
 
 public class RegisterController implements Controller {
