+++ app/src/main/java/com/techcourse/controller/RegisterViewController.java
@@ -2,7 +2,7 @@ package com.techcourse.controller;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.controller.asis.Controller;
+import webmvc.org.springframework.web.servlet.mvc.asis.Controller;
 
 public class RegisterViewController implements Controller {
