+++ build.gradle
@@ -4,6 +4,12 @@ plugins {
     id "org.sonarqube" version "4.2.1.3168"
 }
 
+allprojects {
+    repositories {
+        mavenCentral()
+    }
+}
+
 subprojects {
     apply plugin: 'org.sonarqube'
     sonar {
