+++ mvc/src/main/java/core/org/springframework/util/ReflectionUtils.java
@@ -1,0 +1,38 @@
+package core.org.springframework.util;
+
+import java.lang.reflect.Constructor;
+import java.lang.reflect.Modifier;
+
+public abstract class ReflectionUtils {
+
+    /**
+     * Obtain an accessible constructor for the given class and parameters.
+     * @param clazz the clazz to check
+     * @param parameterTypes the parameter types of the desired constructor
+     * @return the constructor reference
+     * @throws NoSuchMethodException if no such constructor exists
+     * @since 5.0
+     */
+    public static <T> Constructor<T> accessibleConstructor(Class<T> clazz, Class<?>... parameterTypes)
+            throws NoSuchMethodException {
+
+        Constructor<T> ctor = clazz.getDeclaredConstructor(parameterTypes);
+        makeAccessible(ctor);
+        return ctor;
+    }
+
+    /**
+     * Make the given constructor accessible, explicitly setting it accessible
+     * if necessary. The {@code setAccessible(true)} method is only called
+     * when actually necessary, to avoid unnecessary conflicts.
+     * @param ctor the constructor to make accessible
+     * @see Constructor#setAccessible
+     */
+    @SuppressWarnings("deprecation")
+    public static void makeAccessible(Constructor<?> ctor) {
+        if ((!Modifier.isPublic(ctor.getModifiers()) ||
+                !Modifier.isPublic(ctor.getDeclaringClass().getModifiers())) && !ctor.isAccessible()) {
+            ctor.setAccessible(true);
+        }
+    }
+}
