--- mvc/src/main/java/nextstep/mvc/DispatcherServlet.java
@@ -1,51 +1,0 @@
-package nextstep.mvc;
-
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServlet;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.controller.asis.Controller;
-import nextstep.mvc.view.JspView;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-public class DispatcherServlet extends HttpServlet {
-
-    private static final long serialVersionUID = 1L;
-    private static final Logger log = LoggerFactory.getLogger(DispatcherServlet.class);
-
-    private HandlerMapping handlerMappings;
-
-    public void addHandlerMapping(final HandlerMapping handlerMapping) {
-        this.handlerMappings = handlerMapping;
-    }
-
-    @Override
-    public void init() {
-        this.handlerMappings.initialize();
-    }
-
-    @Override
-    protected void service(final HttpServletRequest request, final HttpServletResponse response) throws ServletException {
-        log.debug("Method : {}, Request URI : {}", request.getMethod(), request.getRequestURI());
-
-        try {
-            final var controller = (Controller) handlerMappings.getHandler(request);
-            final var viewName = controller.execute(request, response);
-            move(viewName, request, response);
-        } catch (Throwable e) {
-            log.error("Exception : {}", e.getMessage(), e);
-            throw new ServletException(e.getMessage());
-        }
-    }
-
-    private void move(final String viewName, final HttpServletRequest request, final HttpServletResponse response) throws Exception {
-        if (viewName.startsWith(JspView.REDIRECT_PREFIX)) {
-            response.sendRedirect(viewName.substring(JspView.REDIRECT_PREFIX.length()));
-            return;
-        }
-
-        final var requestDispatcher = request.getRequestDispatcher(viewName);
-        requestDispatcher.forward(request, response);
-    }
-}
