--- mvc/src/main/java/nextstep/mvc/HandlerAdapter.java
@@ -1,12 +1,0 @@
-package nextstep.mvc;
-
-import nextstep.mvc.view.ModelAndView;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-public interface HandlerAdapter {
-    boolean supports(Object handler);
-
-    ModelAndView handle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception;
-}
