--- mvc/src/main/java/nextstep/mvc/HandlerMapping.java
@@ -1,10 +1,0 @@
-package nextstep.mvc;
-
-import jakarta.servlet.http.HttpServletRequest;
-
-public interface HandlerMapping {
-
-    void initialize();
-
-    Object getHandler(HttpServletRequest request);
-}
