--- mvc/src/main/java/nextstep/mvc/controller/tobe/HandlerExecution.java
@@ -1,12 +1,0 @@
-package nextstep.mvc.controller.tobe;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.view.ModelAndView;
-
-public class HandlerExecution {
-
-    public ModelAndView handle(final HttpServletRequest request, final HttpServletResponse response) throws Exception {
-        return null;
-    }
-}
