--- mvc/src/main/java/nextstep/mvc/view/JsonView.java
@@ -1,13 +1,0 @@
-package nextstep.mvc.view;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-import java.util.Map;
-
-public class JsonView implements View {
-
-    @Override
-    public void render(final Map<String, ?> model, final HttpServletRequest request, HttpServletResponse response) throws Exception {
-    }
-}
