--- mvc/src/main/java/nextstep/mvc/view/ModelAndView.java
@@ -1,33 +1,0 @@
-package nextstep.mvc.view;
-
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.Map;
-
-public class ModelAndView {
-
-    private final View view;
-    private final Map<String, Object> model;
-
-    public ModelAndView(final View view) {
-        this.view = view;
-        this.model = new HashMap<>();
-    }
-
-    public ModelAndView addObject(final String attributeName, final Object attributeValue) {
-        model.put(attributeName, attributeValue);
-        return this;
-    }
-
-    public Object getObject(final String attributeName) {
-        return model.get(attributeName);
-    }
-
-    public Map<String, Object> getModel() {
-        return Collections.unmodifiableMap(model);
-    }
-
-    public View getView() {
-        return view;
-    }
-}
