--- mvc/src/main/java/nextstep/mvc/view/View.java
@@ -1,10 +1,0 @@
-package nextstep.mvc.view;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-import java.util.Map;
-
-public interface View {
-    void render(Map<String, ?> model, HttpServletRequest request, HttpServletResponse response) throws Exception;
-}
