--- mvc/src/main/java/nextstep/web/WebApplicationInitializer.java
@@ -1,8 +1,0 @@
-package nextstep.web;
-
-import jakarta.servlet.ServletContext;
-import jakarta.servlet.ServletException;
-
-public interface WebApplicationInitializer {
-    void onStartup(ServletContext servletContext) throws ServletException;
-}
