--- mvc/src/main/java/nextstep/web/annotation/Controller.java
@@ -1,14 +1,0 @@
-package nextstep.web.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target({ElementType.TYPE})
-@Retention(RetentionPolicy.RUNTIME)
-public @interface Controller {
-    String value() default "";
-
-    String path() default "";
-}
