--- mvc/src/main/java/nextstep/web/annotation/RequestMapping.java
@@ -1,16 +1,0 @@
-package nextstep.web.annotation;
-
-import nextstep.web.support.RequestMethod;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target({ElementType.METHOD, ElementType.TYPE})
-@Retention(RetentionPolicy.RUNTIME)
-public @interface RequestMapping {
-    String value() default "";
-
-    RequestMethod[] method() default {};
-}
