--- mvc/src/main/java/nextstep/web/annotation/RequestParam.java
@@ -1,14 +1,0 @@
-package nextstep.web.annotation;
-
-import java.lang.annotation.*;
-
-@Target(ElementType.PARAMETER)
-@Retention(RetentionPolicy.RUNTIME)
-@Documented
-public @interface RequestParam {
-    String value() default "";
-
-    String name() default "";
-
-    boolean required() default true;
-}
