--- mvc/src/main/java/nextstep/web/support/MediaType.java
@@ -1,5 +1,0 @@
-package nextstep.web.support;
-
-public class MediaType {
-    public static final String APPLICATION_JSON_UTF8_VALUE = "application/json;charset=UTF-8";
-}
