--- mvc/src/main/java/nextstep/web/support/RequestMethod.java
@@ -1,5 +1,0 @@
-package nextstep.web.support;
-
-public enum RequestMethod {
-    GET, HEAD, POST, PUT, PATCH, DELETE, OPTIONS, TRACE
-}
