+++ mvc/src/main/java/web/org/springframework/http/MediaType.java
@@ -1,0 +1,5 @@
+package web.org.springframework.http;
+
+public class MediaType {
+    public static final String APPLICATION_JSON_UTF8_VALUE = "application/json;charset=UTF-8";
+}
