+++ mvc/src/main/java/web/org/springframework/web/bind/annotation/PathVariable.java
@@ -1,0 +1,14 @@
+package web.org.springframework.web.bind.annotation;
+
+import java.lang.annotation.*;
+
+@Target(ElementType.PARAMETER)
+@Retention(RetentionPolicy.RUNTIME)
+@Documented
+public @interface PathVariable {
+    String value() default "";
+
+    String name() default "";
+
+    boolean required() default true;
+}
