+++ mvc/src/main/java/web/org/springframework/web/bind/annotation/RequestMapping.java
@@ -1,0 +1,14 @@
+package web.org.springframework.web.bind.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target({ElementType.METHOD, ElementType.TYPE})
+@Retention(RetentionPolicy.RUNTIME)
+public @interface RequestMapping {
+    String value() default "";
+
+    RequestMethod[] method() default {};
+}
