+++ mvc/src/main/java/web/org/springframework/web/bind/annotation/RequestMethod.java
@@ -1,0 +1,5 @@
+package web.org.springframework.web.bind.annotation;
+
+public enum RequestMethod {
+    GET, HEAD, POST, PUT, PATCH, DELETE, OPTIONS, TRACE
+}
