+++ mvc/src/main/java/web/org/springframework/web/bind/annotation/RequestParam.java
@@ -1,0 +1,14 @@
+package web.org.springframework.web.bind.annotation;
+
+import java.lang.annotation.*;
+
+@Target(ElementType.PARAMETER)
+@Retention(RetentionPolicy.RUNTIME)
+@Documented
+public @interface RequestParam {
+    String value() default "";
+
+    String name() default "";
+
+    boolean required() default true;
+}
