+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/tobe/AnnotationHandlerMapping.java
@@ -1,0 +1,29 @@
+package webmvc.org.springframework.web.servlet.mvc.tobe;
+
+import jakarta.servlet.http.HttpServletRequest;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.util.HashMap;
+import java.util.Map;
+
+public class AnnotationHandlerMapping {
+
+    private static final Logger log = LoggerFactory.getLogger(AnnotationHandlerMapping.class);
+
+    private final Object[] basePackage;
+    private final Map<HandlerKey, HandlerExecution> handlerExecutions;
+
+    public AnnotationHandlerMapping(final Object... basePackage) {
+        this.basePackage = basePackage;
+        this.handlerExecutions = new HashMap<>();
+    }
+
+    public void initialize() {
+        log.info("Initialized AnnotationHandlerMapping!");
+    }
+
+    public Object getHandler(final HttpServletRequest request) {
+        return null;
+    }
+}
