+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/tobe/HandlerExecution.java
@@ -1,0 +1,12 @@
+package webmvc.org.springframework.web.servlet.mvc.tobe;
+
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import webmvc.org.springframework.web.servlet.ModelAndView;
+
+public class HandlerExecution {
+
+    public ModelAndView handle(final HttpServletRequest request, final HttpServletResponse response) throws Exception {
+        return null;
+    }
+}
