+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/tobe/HandlerKey.java
@@ -1,0 +1,37 @@
+package webmvc.org.springframework.web.servlet.mvc.tobe;
+
+import web.org.springframework.web.bind.annotation.RequestMethod;
+
+import java.util.Objects;
+
+public class HandlerKey {
+
+    private final String url;
+    private final RequestMethod requestMethod;
+
+    public HandlerKey(final String url, final RequestMethod requestMethod) {
+        this.url = url;
+        this.requestMethod = requestMethod;
+    }
+
+    @Override
+    public String toString() {
+        return "HandlerKey{" +
+                "url='" + url + '\'' +
+                ", requestMethod=" + requestMethod +
+                '}';
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (!(o instanceof HandlerKey)) return false;
+        HandlerKey that = (HandlerKey) o;
+        return Objects.equals(url, that.url) && requestMethod == that.requestMethod;
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(url, requestMethod);
+    }
+}
