+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/view/JsonView.java
@@ -1,0 +1,14 @@
+package webmvc.org.springframework.web.servlet.view;
+
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import webmvc.org.springframework.web.servlet.View;
+
+import java.util.Map;
+
+public class JsonView implements View {
+
+    @Override
+    public void render(final Map<String, ?> model, final HttpServletRequest request, HttpServletResponse response) throws Exception {
+    }
+}
