+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/view/JspView.java
@@ -1,0 +1,31 @@
+package webmvc.org.springframework.web.servlet.view;
+
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import webmvc.org.springframework.web.servlet.View;
+
+import java.util.Map;
+
+public class JspView implements View {
+
+    private static final Logger log = LoggerFactory.getLogger(JspView.class);
+
+    public static final String REDIRECT_PREFIX = "redirect:";
+
+    public JspView(final String viewName) {
+    }
+
+    @Override
+    public void render(final Map<String, ?> model, final HttpServletRequest request, final HttpServletResponse response) throws Exception {
+        // todo
+
+        model.keySet().forEach(key -> {
+            log.debug("attribute name : {}, value : {}", key, model.get(key));
+            request.setAttribute(key, model.get(key));
+        });
+
+        // todo
+    }
+}
