+++ mvc/src/main/resources/META-INF/services/jakarta.servlet.ServletContainerInitializer
@@ -1,1 +1,1 @@
-nextstep.web.NextstepServletContainerInitializer
+web.org.springframework.web.SpringServletContainerInitializer
